requirements.txt
pefile
yara-python
ssdeep-windows
python-magic-bin